+++ src/Finance.Core/BaseLibrary/FoundationTechnologyDevice.cs
@@ -51,7 +51,7 @@ namespace Finance.BaseLibrary
 		/// 
 		/// </summary>
 		[Column("foundation_reliable_hours_id")]
-		public decimal? FoundationReliableHoursId { get; set; }
+		public long? FoundationReliableHoursId { get; set; }
 
 	}
 }
