+++ src/Finance.Core/BaseLibrary/FoundationTechnologyFixture.cs
@@ -43,6 +43,6 @@ namespace Finance.BaseLibrary
 		/// 工序工时id
 		/// </summary>
 		[Column("foundation_reliable_hours_id")]
-		public decimal? FoundationReliableHoursId { get; set; }
+		public long? FoundationReliableHoursId { get; set; }
 	}
 }
