+++ src/Finance.Core/BaseLibrary/FoundationTechnologyFrock.cs
@@ -16,7 +16,7 @@ namespace Finance.BaseLibrary
 		/// 工时工序id
 		/// </summary>
 		[Column("foundation_reliable_hours_id")]
-		public decimal? FoundationReliableHoursId { get; set; }
+		public long? FoundationReliableHoursId { get; set; }
 
 		/// <summary>
 		/// 名称
