+++ src/Finance.Core/BaseLibrary/Logisticscost.cs
@@ -16,7 +16,7 @@ namespace Finance.BaseLibrary
         /// 流程id
         /// </summary>
         [Column("audit_flow_id")]
-		public decimal? AuditFlowId { get; set; }
+		public long? AuditFlowId { get; set; }
 
 		/// <summary>
 		/// 分类
@@ -48,7 +48,7 @@ namespace Finance.BaseLibrary
 		/// 
 		/// </summary>
 		[Column("product_id")]
-		public decimal? ProductId { get; set; }
+		public long? ProductId { get; set; }
 
 		/// <summary>
 		/// 备注
